+++ tests/test_resolve.py
@@ -6,7 +6,7 @@ from unittest.mock import Mock, call
 
 import pytest
 
-from handless import Container, Contextual, ResolutionContext, Singleton, Transient
+from handless import Container, Scope, Scoped, Singleton, Transient
 from handless.lifetimes import Lifetime
 from tests.helpers import FakeService, FakeServiceWithParams
 
@@ -17,7 +17,7 @@ class FactoryRegistrationOptions(TypedDict, total=False):
 
 
 def test_resolve_type_calls_registration_factory_and_returns_its_result(
-    container: Container, context: ResolutionContext
+    container: Container, context: Scope
 ) -> None:
     expected = FakeService()
     factory = Mock(return_value=expected)
@@ -30,7 +30,7 @@ def test_resolve_type_calls_registration_factory_and_returns_its_result(
 
 
 def test_resolve_type_calls_registration_factory_with_ctx_and_returns_its_result(
-    container: Container, context: ResolutionContext
+    container: Container, context: Scope
 ) -> None:
     expected = FakeService()
     factory = Mock(wraps=lambda ctx: expected)  # noqa: ARG005
@@ -43,7 +43,7 @@ def test_resolve_type_calls_registration_factory_with_ctx_and_returns_its_result
 
 
 def test_resolve_type_calls_registration_factory_with_dependencies_and_returns_its_result(
-    container: Container, context: ResolutionContext
+    container: Container, context: Scope
 ) -> None:
     factory = Mock(wraps=FakeServiceWithParams)
     container.register(FakeServiceWithParams).factory(factory)
@@ -60,9 +60,7 @@ def test_resolve_type_calls_registration_factory_with_dependencies_and_returns_i
     "options", [FactoryRegistrationOptions(), FactoryRegistrationOptions(managed=True)]
 )
 def test_resolve_type_enters_context_manager_returned_by_registration_factory(
-    container: Container,
-    context: ResolutionContext,
-    options: FactoryRegistrationOptions,
+    container: Container, context: Scope, options: FactoryRegistrationOptions
 ) -> None:
     container.register(FakeService).self(**options)
 
@@ -73,7 +71,7 @@ def test_resolve_type_enters_context_manager_returned_by_registration_factory(
 
 
 def test_resolve_type_not_enter_context_manager_returned_by_registration_factory_when_managed_is_false(
-    container: Container, context: ResolutionContext
+    container: Container, context: Scope
 ) -> None:
     container.register(FakeService).self(managed=False)
 
@@ -83,7 +81,7 @@ def test_resolve_type_not_enter_context_manager_returned_by_registration_factory
 
 
 def test_resolve_type_not_enter_non_context_manager_object_returned_by_registration_factory(
-    container: Container, context: ResolutionContext
+    container: Container, context: Scope
 ) -> None:
     container.register(object).self(managed=True)
 
@@ -109,7 +107,7 @@ class TestResolveTypeUsingTransientLifetime:
         self,
         request: pytest.FixtureRequest,
         container: Container,
-        context: ResolutionContext,
+        context: Scope,
         factory: Mock,
     ) -> FakeService:
         container.register(FakeService).factory(factory, **request.param)
@@ -117,7 +115,7 @@ class TestResolveTypeUsingTransientLifetime:
         return context.resolve(FakeService)
 
     def test_calls_and_returns_registration_factory_result_on_each_resolve(
-        self, resolved: FakeService, context: ResolutionContext, factory: Mock
+        self, resolved: FakeService, context: Scope, factory: Mock
     ) -> None:
         received = context.resolve(FakeService)
 
@@ -125,20 +123,16 @@ class TestResolveTypeUsingTransientLifetime:
         factory.assert_has_calls([call(_=context), call(_=context)])
 
     def test_calls_and_returns_registration_factory_result_on_different_context(
-        self,
-        resolved: FakeService,
-        container: Container,
-        context: ResolutionContext,
-        factory: Mock,
+        self, resolved: FakeService, container: Container, context: Scope, factory: Mock
     ) -> None:
-        with container.open_context() as context2:
+        with container.create_scope() as context2:
             received = context2.resolve(FakeService)
 
         assert received is not resolved
         factory.assert_has_calls([call(_=context), call(_=context2)])
 
     def test_release_context_exit_entered_context_manager(
-        self, context: ResolutionContext, resolved: FakeService
+        self, context: Scope, resolved: FakeService
     ) -> None:
         another = context.resolve(FakeService)
 
@@ -155,14 +149,14 @@ class TestResolveTypeBoundToSingletonRegistration:
 
     @pytest.fixture
     def resolved(
-        self, container: Container, context: ResolutionContext, factory: Mock
+        self, container: Container, context: Scope, factory: Mock
     ) -> FakeService:
         container.register(FakeService).factory(factory, Singleton())
 
         return context.resolve(FakeService)
 
     def test_calls_and_returns_registration_factory_result_once_per_context(
-        self, resolved: FakeService, context: ResolutionContext, factory: Mock
+        self, resolved: FakeService, context: Scope, factory: Mock
     ) -> None:
         received = context.resolve(FakeService)
 
@@ -176,14 +170,14 @@ class TestResolveTypeBoundToSingletonRegistration:
         context: Container,
         factory: Mock,
     ) -> None:
-        with container.open_context() as context2:
+        with container.create_scope() as context2:
             received = context2.resolve(FakeService)
 
         assert received is resolved
         factory.assert_called_once_with(_=context)
 
     def test_resolve_singleton_is_threadsafe(self, container: Container) -> None:
-        def _factory(ctx: ResolutionContext) -> FakeServiceWithParams:
+        def _factory(ctx: Scope) -> FakeServiceWithParams:
             # Small sleep to force threads context switch
             time.sleep(0.01)
             return FakeServiceWithParams(ctx.resolve(str), ctx.resolve(int))
@@ -195,14 +189,14 @@ class TestResolveTypeBoundToSingletonRegistration:
 
         with ThreadPoolExecutor(100) as pool:
             results = pool.map(
-                lambda _: container.open_context().resolve(FakeService), range(100)
+                lambda _: container.create_scope().resolve(FakeService), range(100)
             )
 
         mock.assert_called_once()
         assert len(set(results)) == 1
 
     def test_release_context_not_exit_entered_context_manager(
-        self, context: ResolutionContext, resolved: FakeService
+        self, context: Scope, resolved: FakeService
     ) -> None:
         context.release()
 
@@ -216,7 +210,7 @@ class TestResolveTypeBoundToSingletonRegistration:
         assert resolved.exited
 
     def test_release_container_clear_cached_value(
-        self, container: Container, context: ResolutionContext, resolved: FakeService
+        self, container: Container, context: Scope, resolved: FakeService
     ) -> None:
         container.release()
 
@@ -225,7 +219,7 @@ class TestResolveTypeBoundToSingletonRegistration:
         assert received is not resolved
 
     def test_release_context_not_clear_cached_value(
-        self, context: ResolutionContext, resolved: FakeService
+        self, context: Scope, resolved: FakeService
     ) -> None:
         context.release()
 
@@ -241,14 +235,14 @@ class TestResolveTypeBoundToContextRegistration:
 
     @pytest.fixture(autouse=True)
     def resolved(
-        self, container: Container, context: ResolutionContext, factory: Mock
+        self, container: Container, context: Scope, factory: Mock
     ) -> FakeService:
-        container.register(FakeService).factory(factory, Contextual())
+        container.register(FakeService).factory(factory, Scoped())
 
         return context.resolve(FakeService)
 
     def test_calls_and_returns_registration_factory_result_once_per_context(
-        self, resolved: FakeService, context: ResolutionContext, factory: Mock
+        self, resolved: FakeService, context: Scope, factory: Mock
     ) -> None:
         received = context.resolve(FakeService)
 
@@ -256,27 +250,23 @@ class TestResolveTypeBoundToContextRegistration:
         factory.assert_called_once_with(_=context)
 
     def test_calls_and_returns_registration_factory_result_on_different_context(
-        self,
-        resolved: FakeService,
-        container: Container,
-        context: ResolutionContext,
-        factory: Mock,
+        self, resolved: FakeService, container: Container, context: Scope, factory: Mock
     ) -> None:
-        with container.open_context() as context2:
+        with container.create_scope() as context2:
             received = context2.resolve(FakeService)
 
         assert received is not resolved
         factory.assert_has_calls([call(_=context), call(_=context2)])
 
     def test_release_context_exit_entered_context_manager(
-        self, context: ResolutionContext, resolved: FakeService
+        self, context: Scope, resolved: FakeService
     ) -> None:
         context.release()
 
         assert resolved.exited
 
     def test_release_context_clear_cached_value(
-        self, context: ResolutionContext, resolved: FakeService
+        self, context: Scope, resolved: FakeService
     ) -> None:
         context.release()
 
@@ -299,7 +289,7 @@ class TestOverrideTypes:
         return factory_override
 
     def test_resolve_type_calls_override_factory_and_returns_its_result_when_registered(
-        self, context: ResolutionContext, factory: Mock, factory_override: Mock
+        self, context: Scope, factory: Mock, factory_override: Mock
     ) -> None:
         resolved = context.resolve(FakeService)
 
@@ -310,7 +300,7 @@ class TestOverrideTypes:
     def test_release_container_clear_overrides(
         self,
         container: Container,
-        context: ResolutionContext,
+        context: Scope,
         factory: Mock,
         factory_override: Mock,
     ) -> None:
@@ -323,7 +313,7 @@ class TestOverrideTypes:
     def test_override_can_override_an_already_overridden_type(
         self,
         container: Container,
-        context: ResolutionContext,
+        context: Scope,
         factory: Mock,
         factory_override: Mock,
     ) -> None:
@@ -340,16 +330,16 @@ class TestOverrideTypes:
         ("lifetime", "override_lifetime"),
         [
             # Ensure that whatever lifetimes are used, override always takes precedence
-            *itertools.permutations([Transient(), Contextual(), Singleton()], 2),
+            *itertools.permutations([Transient(), Scoped(), Singleton()], 2),
             (Transient(), Transient()),
             (Singleton(), Singleton()),
-            (Contextual(), Contextual()),
+            (Scoped(), Scoped()),
         ],
     )
     def test_override_a_cached_type_returns_override_result(
         self,
         container: Container,
-        context: ResolutionContext,
+        context: Scope,
         lifetime: Lifetime,
         override_lifetime: Lifetime,
     ) -> None: